Editar email admin, teacher, parent, student
Cuando se edita el email se verifica que no existe ya en la tabla ni sea el del mismo id

+++ application/models/Admin_model.php
@@ -10,21 +10,37 @@ class Admin_model extends CI_Model {
 
 
     function createNewAdministrator(){
-
-        $page_data['name']  = html_escape($this->input->post('name'));
-        $page_data['email']  = html_escape($this->input->post('email'));
-        $page_data['phone']  = html_escape($this->input->post('phone'));
-        $page_data['password']  = sha1($this->input->post('name'));
-        $page_data['level']     = html_escape($this->input->post('level'));
-        $this->db->insert('admin', $page_data);
-        $admin_id = $this->db->insert_id();
-        move_uploaded_file($_FILES['admin_image']['tmp_name'], 'uploads/admin_image/' . $admin_id . '.jpg');
-
-        $page_data2['admin_id'] =  $admin_id;
-        $this->db->insert('admin_role', $page_data2);
-
-
+        $name = html_escape($this->input->post('name'));
+        $email = html_escape($this->input->post('email'));
+        $phone = html_escape($this->input->post('phone'));
+        $password = sha1($this->input->post('password'));
+        $level = html_escape($this->input->post('level'));
+    
+        // Verificar si el correo electrónico ya existe en la tabla "admin"
+        $existing_admin = $this->db->get_where('admin', array('email' => $email))->row();
+    
+        if ($existing_admin) {
+            // Si ya existe un administrador con el mismo correo electrónico, muestra un mensaje de error
+            $this->session->set_flashdata('error_message', 'El correo electrónico ya esta registrado');
+            redirect(base_url(). 'admin/newAdministrator', 'refresh'); // Cambia 'admin/some_page' a la URL deseada
+        } else {
+            // Insertar el nuevo administrador si no hay duplicados
+            $page_data = array(
+                'name' => $name,
+                'email' => $email,
+                'phone' => $phone,
+                'password' => $password,
+                'level' => $level
+            );
+            $this->db->insert('admin', $page_data);
+            $admin_id = $this->db->insert_id();
+            move_uploaded_file($_FILES['admin_image']['tmp_name'], 'uploads/admin_image/' . $admin_id . '.jpg');
+    
+            $page_data2['admin_id'] = $admin_id;
+            $this->db->insert('admin_role', $page_data2);
+        }
     }
+    
 
     function deleteAdministrator($param2){
